requirements.txt
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyyaml==6.0.1
scipy==1.11.4
scikit-learn==1.3.2
jupyter==1.0.0
ipython==8.18.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
tqdm==4.66.1
black==23.12.0
flake8==6.1.0